Cleanup and add Cache-Control for FileHTTPHandler
