+++ FlyingFox/Sources/HTTPHeader.swift
@@ -76,60 +76,3 @@ public extension HTTPHeader {
     static let webSocketVersion    = HTTPHeader("Sec-WebSocket-Version")
     static let xForwardedFor       = HTTPHeader("X-Forwarded-For")
 }
-
-
-public enum CacheControl {
-    public enum ResponseDirective: Sendable, CustomStringConvertible {
-        case maxAge(Int)
-        case sharedMaxAge(Int)
-        case noCache
-        case noStore
-        case noTransform
-        case mustRevalidate
-        case proxyRevalidate
-        case mustUnderstand
-        case `private`
-        case `public`
-        case immutable
-        case staleWhileRevalidate
-        case staleIfError
-        
-        public var description: String {
-            switch self {
-            case .maxAge(let value):
-                return "max-age=\(value)"
-            case .sharedMaxAge(let value):
-                return "s-maxage=\(value)"
-            case .noCache:
-                return "no-cache"
-            case .noStore:
-                return "no-store"
-            case .noTransform:
-                return "no-transform"
-            case .mustRevalidate:
-                return "must-revalidate"
-            case .proxyRevalidate:
-                return "proxy-revalidate"
-            case .mustUnderstand:
-                return "must-understand"
-            case .private:
-                return "private"
-            case .public:
-                return "public"
-            case .immutable:
-                return "immutable"
-            case .staleWhileRevalidate:
-                return "stale-while-revalidate"
-            case .staleIfError:
-                return "stale-if-error"
-            }
-        }
-    }
-}
-
-extension [CacheControl.ResponseDirective] {
-    func serialized() -> String {
-        let directives: [CacheControl.ResponseDirective] = self.isEmpty ? [.private] : self
-        return Set(directives.map({ $0.description })).joined(separator: ",")
-    }
-}
