+++ FlyingFox/Sources/Handlers/DirectoryHTTPHandler.swift
@@ -30,9 +30,6 @@
 //
 
 import Foundation
-#if canImport(CryptoKit)
-import CryptoKit
-#endif
 
 public struct DirectoryHTTPHandler: HTTPHandler {
 
@@ -61,31 +58,28 @@ public struct DirectoryHTTPHandler: HTTPHandler {
             let data = try? Data(contentsOf: filePath) else {
             return HTTPResponse(statusCode: .notFound)
         }
-        
+
         var headers: HTTPHeaders = [
             .contentType: FileHTTPHandler.makeContentType(for: filePath.absoluteString),
             .cacheControl: cacheControl.serialized(),
-            .date: Self.dateFormatter.string(from: Date())
+            .date: CacheControl.getDateHeaderValue()
         ]
-        
-        if let expiresValue = generateExpiresValue(for: filePath) {
+
+        if let expiresValue = CacheControl.generateExpiresValue(for: filePath) {
             headers[.lastModified] = expiresValue
-            
             if let ifModifiedSince = request.headers[.ifModifiedSince], expiresValue == ifModifiedSince {
                 return HTTPResponse(statusCode: .notModified,
                                     headers: headers)
             }
         }
-        
-#if canImport(CryptoKit)
-        let eTag = generateETag(for: data)
-        headers[.eTag] = eTag
-        
-        if let ifNoneMatch = request.headers[.ifNoneMatch], eTag == ifNoneMatch {
-            return HTTPResponse(statusCode: .notModified,
-                                headers: headers)
+
+        if let eTag = CacheControl.generateETagValue(for: data) {
+            headers[.eTag] = eTag
+            if let ifNoneMatch = request.headers[.ifNoneMatch], eTag == ifNoneMatch {
+                return HTTPResponse(statusCode: .notModified,
+                                    headers: headers)
+            }
         }
-#endif
 
         return HTTPResponse(
             statusCode: .ok,
@@ -108,37 +102,4 @@ public struct DirectoryHTTPHandler: HTTPHandler {
         return root?.appendingPathComponent(subPath)
     }
     
-    private static let dateFormatter: DateFormatter = {
-        let df = DateFormatter()
-        df.dateFormat = "EEE, d MMM yyyy HH:mm:ss zzz"
-        df.timeZone = TimeZone(secondsFromGMT: 0)
-        df.locale = Locale(identifier: "en_US_POSIX")
-        return df
-    }()
-    
-    private func generateExpiresValue(for filePath: URL) -> String? {
-        do {
-            let path = {
-                if #available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *) {
-                    return filePath.path()
-                } else {
-                    return filePath.path
-                }
-            }()
-            let attributes = try FileManager.default.attributesOfItem(atPath: path)
-            if let modificationDate = attributes[FileAttributeKey.modificationDate] as? Date ?? attributes[FileAttributeKey.creationDate] as? Date {
-                return Self.dateFormatter.string(from: modificationDate)
-            }
-        } catch {
-        }
-        return nil
-    }
-    
-#if canImport(CryptoKit)
-    private func generateETag(for data: Data) -> String {
-        let sha256digest = SHA256.hash(data: data)
-        let eTag = "\"\(sha256digest.map { String(format: "%02x", $0) }.joined())\""
-        return eTag
-    }
-#endif
 }
